+++ src/main/java/me/marcosavarino/clinica_odontologica/service/impl/PacienteService.java
@@ -3,10 +3,17 @@ package me.marcosavarino.clinica_odontologica.service.impl;
 import me.marcosavarino.clinica_odontologica.dto.response.*;
 import me.marcosavarino.clinica_odontologica.dto.response.ResponsesTurno.Pacientes.PacienteTurnoResponseDto;
 import me.marcosavarino.clinica_odontologica.dto.response.ResponsesTurno.Pacientes.TurnosPacienteResponseDto;
+import me.marcosavarino.clinica_odontologica.entity.Domicilio;
 import me.marcosavarino.clinica_odontologica.entity.Paciente;
+import me.marcosavarino.clinica_odontologica.exception.BadRequestException;
+import me.marcosavarino.clinica_odontologica.exception.GlobalHandler;
+import me.marcosavarino.clinica_odontologica.exception.ResourceNotFoundException;
 import me.marcosavarino.clinica_odontologica.repository.IPacienteRepository;
 import me.marcosavarino.clinica_odontologica.service.IPacienteService;
+import me.marcosavarino.clinica_odontologica.utils.ValidationUtils;
 import org.modelmapper.ModelMapper;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -18,6 +25,7 @@ import java.util.stream.Collectors;
 @Service
 public class PacienteService implements IPacienteService {
     private IPacienteRepository pacienteRepository;
+    private final Logger logger = LoggerFactory.getLogger(PacienteService.class);
 
     @Autowired
     private ModelMapper modelMapper;
@@ -28,19 +36,26 @@ public class PacienteService implements IPacienteService {
 
     @Override
     public Paciente guardarPaciente(Paciente paciente) {
+        ValidationUtils.validatePaciente(paciente);
+        ValidationUtils.validateDomicilio(paciente.getDomicilio());
+        logger.info("Guardando paciente: {}", paciente);
         return pacienteRepository.save(paciente);
     }
 
     @Override
     public Optional<Paciente> buscarPorId(Integer id) {
-        return pacienteRepository.findById(id);
+        Optional<Paciente> paciente = pacienteRepository.findById(id);
+        if (!paciente.isPresent()) {
+            throw new BadRequestException("No se encontro al paciente con el " + id);
+        }
+        return paciente;
     }
 
     @Override
     public List<PacienteTurnoResponseDto> buscarTodosLosPacientes() {
         List<Paciente> pacientesList = pacienteRepository.findAll();
         List<PacienteTurnoResponseDto> pacienteDto = new ArrayList<>();
-        for (Paciente paciente : pacientesList){
+        for (Paciente paciente : pacientesList) {
             PacienteTurnoResponseDto pacienteAuxiliar = mapearPaciente(paciente);
             pacienteDto.add(pacienteAuxiliar);
         }
@@ -49,12 +64,26 @@ public class PacienteService implements IPacienteService {
 
     @Override
     public void pacienteUpdate(Paciente p) {
-        pacienteRepository.save(p);
+        Optional<Paciente> paciente = pacienteRepository.findById(p.getId());
+        if (paciente.isPresent()) {
+            ValidationUtils.validatePaciente(p);
+            ValidationUtils.validateDomicilio(p.getDomicilio());
+            pacienteRepository.save(p);
+        } else {
+            logger.warn("No se encontró el paciente con ID {}", p.getId());
+            throw new BadRequestException("No se encontró al paciente con ID " + p.getId());
+        }
+
     }
 
     @Override
     public void pacienteDelete(Integer id) {
-        pacienteRepository.deleteById(id);
+        Optional<Paciente> paciente = pacienteRepository.findById(id);
+        if (paciente.isPresent()) {
+            pacienteRepository.deleteById(id);
+        } else {
+            throw new ResourceNotFoundException("No se pudo eliminar el odontologo " + id);
+        }
     }
 
     @Override
@@ -64,7 +93,7 @@ public class PacienteService implements IPacienteService {
             PacienteTurnoResponseDto responseDto = mapearPaciente(pacienteOpt.get());
             return Optional.of(responseDto);
         } else {
-            return Optional.empty();
+            throw new BadRequestException("No se encontro al paciente con el " + id);
         }
     }
